AlloyML/requirements.txt
absl-py==0.11.0
astunparse==1.6.3
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
numexpr==2.7.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandastable==0.12.2.post1
Pillow==8.1.0
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygad==2.12.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
